Completely remove isomorphic-fetch dep and just global fetch
optionalDependencies still installs in depending projects. Just fall back to global browser fetch, it'll work in browsers and it'll still work in our tests because we inject our own fetch.

It'll fail in Node.js if they try to use the default requester, but they can just supply their own.